fix(tracking): widen token-count aggregates to long, matching Rust's 64-bit usize accumulation
Rust's Tracker accumulates input/output/saved token sums into usize
(64-bit on all supported targets), but the C# port narrowed the
equivalent SUM(...) query results to int via (int) casts on
reader.GetInt64(...). A busy tracking window's SUM(saved_tokens) could
plausibly exceed int.MaxValue (~2.1B), silently wrapping in C# where
Rust would not.

Widened GainSummary.TotalInput/TotalOutput/TotalSaved,
DayStats/WeekStats/MonthStats.InputTokens/OutputTokens/SavedTokens,
CommandStat.SavedTokens, CommandRecord.SavedTokens, and
DaySavings.SavedTokens to long, removing the narrowing (int) casts in
Tracker.cs. Row-count fields (TotalCommands, Commands, Count) are left
as int — Rust types those usize fields the same way it types the token
sums, but they're bounded by row counts that don't carry the same
overflow risk the review flagged, and Rust's own tests don't exercise
them past int range either.

+++ RtkSharp/Core/Tracking/Models.cs
@@ -16,7 +16,7 @@ public sealed class CommandRecord
     public required string RtkCmd { get; init; }
 
     /// <summary>Number of tokens saved (input minus output).</summary>
-    public required int SavedTokens { get; init; }
+    public required long SavedTokens { get; init; }
 
     /// <summary>Savings percentage (<c>(saved / input) * 100</c>).</summary>
     public required double SavingsPct { get; init; }
@@ -36,7 +36,7 @@ public sealed class CommandStat
     public required int Count { get; init; }
 
     /// <summary>Total tokens saved across all invocations of this command.</summary>
-    public required int SavedTokens { get; init; }
+    public required long SavedTokens { get; init; }
 
     /// <summary>Average savings percentage across all invocations of this command.</summary>
     public required double AvgSavingsPct { get; init; }
@@ -55,7 +55,7 @@ public sealed class DaySavings
     public required string Date { get; init; }
 
     /// <summary>Total tokens saved on this day.</summary>
-    public required int SavedTokens { get; init; }
+    public required long SavedTokens { get; init; }
 }
 
 /// <summary>
@@ -69,13 +69,13 @@ public sealed class GainSummary
     public required int TotalCommands { get; init; }
 
     /// <summary>Total input tokens across all commands.</summary>
-    public required int TotalInput { get; init; }
+    public required long TotalInput { get; init; }
 
     /// <summary>Total output tokens across all commands.</summary>
-    public required int TotalOutput { get; init; }
+    public required long TotalOutput { get; init; }
 
     /// <summary>Total tokens saved (input minus output) across all commands.</summary>
-    public required int TotalSaved { get; init; }
+    public required long TotalSaved { get; init; }
 
     /// <summary>Average savings percentage across all commands.</summary>
     public required double AvgSavingsPct { get; init; }
@@ -106,13 +106,13 @@ public sealed class DayStats
     public required int Commands { get; init; }
 
     /// <summary>Total input tokens for this day.</summary>
-    public required int InputTokens { get; init; }
+    public required long InputTokens { get; init; }
 
     /// <summary>Total output tokens for this day.</summary>
-    public required int OutputTokens { get; init; }
+    public required long OutputTokens { get; init; }
 
     /// <summary>Total tokens saved this day.</summary>
-    public required int SavedTokens { get; init; }
+    public required long SavedTokens { get; init; }
 
     /// <summary>Savings percentage for this day.</summary>
     public required double SavingsPct { get; init; }
@@ -142,13 +142,13 @@ public sealed class WeekStats
     public required int Commands { get; init; }
 
     /// <summary>Total input tokens for this week.</summary>
-    public required int InputTokens { get; init; }
+    public required long InputTokens { get; init; }
 
     /// <summary>Total output tokens for this week.</summary>
-    public required int OutputTokens { get; init; }
+    public required long OutputTokens { get; init; }
 
     /// <summary>Total tokens saved this week.</summary>
-    public required int SavedTokens { get; init; }
+    public required long SavedTokens { get; init; }
 
     /// <summary>Savings percentage for this week.</summary>
     public required double SavingsPct { get; init; }
@@ -173,13 +173,13 @@ public sealed class MonthStats
     public required int Commands { get; init; }
 
     /// <summary>Total input tokens for this month.</summary>
-    public required int InputTokens { get; init; }
+    public required long InputTokens { get; init; }
 
     /// <summary>Total output tokens for this month.</summary>
-    public required int OutputTokens { get; init; }
+    public required long OutputTokens { get; init; }
 
     /// <summary>Total tokens saved this month.</summary>
-    public required int SavedTokens { get; init; }
+    public required long SavedTokens { get; init; }
 
     /// <summary>Savings percentage for this month.</summary>
     public required double SavingsPct { get; init; }
